hive connector fixes implicit cast of source type Long to narrower destination type int

+++ inlong-sort/sort-flink/sort-flink-v1.13/sort-connectors/hive/src/main/java/org/apache/inlong/sort/hive/HiveRowDataPartitionComputer.java
@@ -164,7 +164,7 @@ public class HiveRowDataPartitionComputer extends RowDataPartitionComputer {
 
                 boolean replaceLineBreak = hiveWriterFactory.getStorageDescriptor().getInputFormat()
                         .contains("TextInputFormat");
-                Pair<GenericRowData, Long> pair = HiveTableUtil.getRowData(rawData, columnNames, allTypes,
+                Pair<GenericRowData, Integer> pair = HiveTableUtil.getRowData(rawData, columnNames, allTypes,
                         replaceLineBreak);
                 GenericRowData genericRowData = pair.getLeft();
